refactor(sysmon): move sysmon components to src/workspaces/sysmon/
