+++ src/workspaces/sysmon/DashboardEventTypeChart.tsx
@@ -1,6 +1,6 @@
 import { tokens } from "@fluentui/react-components";
 import { DonutChart } from "@fluentui/react-charts";
-import type { SysmonSummary } from "../../workspaces/sysmon/types";
+import type { SysmonSummary } from "./types";
 
 interface DashboardEventTypeChartProps {
   summary: SysmonSummary;
