+++ src/workspaces/sysmon/SysmonConfigView.tsx
@@ -1,5 +1,5 @@
 import { tokens } from "@fluentui/react-components";
-import { useSysmonStore } from "../../workspaces/sysmon/sysmon-store";
+import { useSysmonStore } from "./sysmon-store";
 import { LOG_MONOSPACE_FONT_FAMILY } from "../../lib/log-accessibility";
 
 export function SysmonConfigView() {
